Make holmes python installable (#57)
Move contents of holmes.py to holmes/main.py. This is necesssary to make
the entrypoint installable.

Note: this includes changes from #56 as well, so will be easier to
review once that is merged. (I didn't base this PR off master as that
would have caused conflicts once #56 is merged as they both touch
holmes.py.)