Increase accuracy of `holmes ask` results (#89)
Ready to merge. 

Test results are here -
https://docs.google.com/document/d/1o9hiAcuIhPrtHex7IupWVChjS4FykNj_Vxmv2ZJfHJ0/
(note: this is a private internal robusta-link as our test framework is
currently not open source)

This PR increases investigation accuracy, as well as bringing the
generic_investigation and generic_ask templates closer to one another,
which makes them slightly easier to maintain.