Add kafka tool (#229)
Add kafka tool to Holmes.
Supported authentication is only username/password or anonymous.

This PR also makes some unrelated tweaks to the code in master and adds
mocks to the `ask_holmes:21_job_fail_curl_no_svc_account` test case so
that it passed without the need to run it live

---------

Co-authored-by: avi@robusta.dev <avi@robusta.dev>
Co-authored-by: Avi-Robusta <97387909+Avi-Robusta@users.noreply.github.com>