ROB-1022 coralogix toolset returns url (#366)
Some improvements to coralogix:

- The log tool returns 
  - A URL to replicate the query on the coralogix UI
  - The dataprime query
- The log level param has been removed. I'm not sure Holmes would use it
well today and other logs toolsets don't support it. We may introduce it
later if we find that Holmes need to quickly narrow down on errors
although that may be solved better with an additional grep-like filter
as opposed to log-level.
- The default time window for logs is extended to 24hrs. This is to help
Holmes get logs when investigating pod related issues and the pod has
been deleted.

:warning: should be merged after
https://github.com/robusta-dev/holmesgpt/pull/365 because this branch
was forked off of #365.