DONGMERGE chore: allow customer to specify the AKS env in aks toolset (#404)
Normally, a user has to specify the cluster name, resource group and
subscription ID to locate an AKS cluster. Considering it's not possible
to extract these AKS info from the Kubernetes resources, To store the
context on the cluster info, we require the user to set these
environment RESOURCE_GROUP_NAME, CLUSTER_NAME and SUBSCRIPTION_ID to
load aks toolset. This saves the effort to provide the same information
every time they want to debug an AKS cluster.

Signed-off-by: Qingchuan Hao <qingchuan.hao@microsoft.com>
Co-authored-by: arik <alon.arik@gmail.com>