chore: add issue template for bug and feature (#420)
To simplify the process of creating issues, we have added issue
templates for bug and feature requests. The templates are basically
copied from
https://github.com/kubernetes/kubernetes/tree/master/.github/ISSUE_TEMPLATE
but removed the items not related to homlesgpt. This is the initial
attempt to add the template, and we can see how to add details into it
to better describe a bug like introducing AI backend, and toolset type,
and what local environment is necessary for us to debug the issue.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a structured bug report template to streamline and improve
bug submissions.
- Added an enhancement tracking template for submitting and detailing
feature requests.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Signed-off-by: Qingchuan Hao <qingchuan.hao@microsoft.com>