Fix version warning from pipx (#544)
Fixes #466 
Note: this is only a partial fix, as now holmes will disable version
checks altogether when installed from a github source version.

I think this is OK, as once we push to pypi users will be instructed to
install from pypi and the version check will work for them.