feat(runbook): introduce runbook toolset to fetch internal runbooks (#547)
This PR introduces a toolset Runbook to fetch the built-in runbooks, the
catalog are integrated into the system prompt. I closed the PR to let
llm pick the runbook and append it to the initial steps before executing
any tools https://github.com/robusta-dev/holmesgpt/pull/533, but it's
not already necessary to follow a runbook and the runbook selected might
be misleading depending only on the initial user prompt.
The result of toolset will not be appended to the user prompt, but as
the other tool result, it will be part of the assistant prompt