add test case (#580)
This test was originally added as one of Holmes users had issues were
holmes used the kubectl -A tools and not kubectl -n thus getting a lot
of pods and some of the data got truncated.

However, already holmes passes this scenario on latest version.

I'm adding the evals for the future as it was already written.

---------

Co-authored-by: Nicolas Herment <nherment@users.noreply.github.com>