Dont mark failing evals as passed (#619)
Today, when correctness=0 evals are marked as passing even when they
actually failed. This makes it very hard to understand what is actually
passing or failing on a run using only pytest output.

This PR allows us to clearly mark those evals as failed, while not
failing the github check because they are not a regression.