Small eval improvements + investigation-refusal eval (#966)
Add an eval to test for undesired behaviour reported by a customer:
holmes refuses to investigate kafka lag using Prometheus, because it
insists on telling the user to enable the Kafka toolset first.

This behaviour will be fixed in a separate PR - this PR only includes
the eval.
This PR also improves error handling in the evals framework.